refactor: なんだか忘れた

+++ src/lib/mypow.c
@@ -1,5 +1,5 @@
-short mypow(short base, short exponent) {
-  short output = 1;
+int mypow(short base, short exponent) {
+  int output = 1;
 
   while(exponent != 0) {
     output = output * base;
